requirements.txt
python-telegram-bot==13.7
openai==0.28.0
psycopg2-binary==2.9.5
python-dotenv==0.21.0
